change: Move scripts into the core
