+++ src/peltak/core/scripts/__init__.py
