+++ test/unit/extra/scripts/commands/test_post_conf_load.py
@@ -1,7 +1,7 @@
 # pylint: disable=missing-docstring
 from unittest.mock import Mock, patch
 
-from peltak.extra.scripts.commands import post_conf_load, root_cli
+from peltak.commands.scripts import post_conf_load, root_cli
 from peltak.testing import patch_pelconf
 
 
@@ -11,7 +11,7 @@ from peltak.testing import patch_pelconf
         'fake2': {'command': 'fake2_cmd'},
     }
 })
-@patch('peltak.extra.scripts.commands.Script')
+@patch('peltak.core.scripts.types.Script')
 def test_works(p_Script):
     p_script_1 = Mock()
     p_script_2 = Mock()
@@ -31,7 +31,7 @@ def test_works(p_Script):
         'fake1': {'command': 'fake1_cmd', 'root_cli': True},
     }
 })
-@patch('peltak.extra.scripts.commands.Script')
+@patch('peltak.core.scripts.types.Script')
 def test_attaches_to_root_cli_if_root_cli_is_True(p_Script):
     p_script = Mock()
     p_script.root_cli = True
